requirements.txt
fastapi
uvicorn[standard]
transformers
torch
Pillow
bitsandbytes
accelerate




